backend/requirements.txt
tqdm
pyyaml
numpy==1.16.3
pandas==1.1.0
matplotlib==3.1.0
pillow
opencv-python
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.2.0+cpu
torchvision==0.4.0+cpu